requirements.txt
hf_transfer
dotenv
torch==2.6
torchvision==0.21.0
transformers
accelerate==0.26.0
sentence-transformers
pandas
pydantic
datasets
Pillow
jinja2==3.1.0
tokenizers
cragmm-search-pipeline>=0.5.0
chromadb>=1.0.0
loguru
rich
tqdm
huggingface-hub